모의고사 문제 CRUD 기능 추가

+++ src/main/java/Project/OpenBook/Dto/chapter/ChapterContentDto.java
@@ -13,5 +13,5 @@ public class ChapterContentDto {
     private Integer number;
     private Integer startDate;
     private Integer endDate;
-    private Long topicCount;
+    private int topicCount;
 }
